Added transformation to multiple formats
