+++ cvat/apps/dataset_manager/formats/icdar.py
@@ -14,6 +14,7 @@ from cvat.apps.dataset_manager.bindings import (GetCVATDataExtractor,
     import_dm_annotations)
 from cvat.apps.dataset_manager.util import make_zip_archive
 
+from .transformations import RotatedBoxesToPolygons
 from .registry import dm_env, exporter, importer
 
 
@@ -116,6 +117,7 @@ def _export_segmentation(dst_file, instance_data, save_images=False):
     dataset = Dataset.from_extractors(GetCVATDataExtractor(
         instance_data, include_images=save_images), env=dm_env)
     with TemporaryDirectory() as temp_dir:
+        dataset.transform(RotatedBoxesToPolygons)
         dataset.transform('polygons_to_masks')
         dataset.transform('boxes_to_masks')
         dataset.transform('merge_instance_segments')
